+++ tests/config_complete_defaults.json
@@ -1,0 +1,76 @@
+{
+    "stepFile": "tests/cuboid.stp",
+    "Options": {
+        "forceCylinder": false, 
+        "newSplitPlane": true, 
+        "delLastNumber": false, 
+        "enlargeBox": 2.0, 
+        "nPlaneReverse": 0, 
+        "splitTolerance": 0.0, 
+        "scaleUp": true, 
+        "quadricPY": false, 
+        "Facets": false, 
+        "prnt3PPlane": false, 
+        "forceNoOverlap": false
+    },
+    "Tolerances": {
+        "relativeTol": false, 
+        "relativePrecision": 1e-06, 
+        "value": 1e-06, 
+        "distance": 0.0001, 
+        "angle": 0.0001, 
+        "pln_distance": 0.0001, 
+        "pln_angle": 0.0001, 
+        "cyl_distance": 0.0001, 
+        "cyl_angle": 0.0001, 
+        "sph_distance": 0.0001, 
+        "kne_distance": 0.0001, 
+        "kne_angle": 0.0001, 
+        "tor_distance": 0.0001, 
+        "tor_angle": 0.0001, 
+        "min_area": 0.01
+    },
+    "NumericFormat": {
+        "P_abc": "14.7e", 
+        "P_d": "14.7e", 
+        "P_xyz": "14.7e", 
+        "S_r": "14.7e", 
+        "S_xyz": "14.7e", 
+        "C_r": "12f", 
+        "C_xyz": "12f", 
+        "K_xyz": "13.6e", 
+        "K_tan2": "12f", 
+        "T_r": "14.7e", 
+        "T_xyz": "14.7e", 
+        "GQ_1to6": "18.15f", 
+        "GQ_7to9": "18.15f", 
+        "GQ_10": "18.15f"
+    },
+    "Settings": {
+        "matFile": "", 
+        "voidGen": true, 
+        "debug": false, 
+        "compSolids": true, 
+        "simplify": "no", 
+        "cellRange": [], 
+        "exportSolids": "", 
+        "minVoidSize": 200.0, 
+        "maxSurf": 50, 
+        "maxBracket": 30, 
+        "voidMat": [], 
+        "voidExclude": [], 
+        "startCell": 1, 
+        "startSurf": 1, 
+        "sort_enclosure": false
+    },
+    "export_csg":{
+        "title": "Converted with GEOUNED",
+        "geometryName": "csg",
+        "outFormat": ["openMC_XML", "openMC_PY", "serpent", "phits", "mcnp"],
+        "volSDEF": false,
+        "volCARD": true,
+        "dummyMat": false,
+        "cellCommentFile": false,
+        "cellSummaryFile": true
+    }
+}
